requirements.txt
# Core Machine Learning & NLP Libraries
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
transformers==4.48.3
numpy~=2.0.2

# FastAPI Libraries
uvicorn~=0.34.0
fastapi~=0.115.6
pydantic~=2.10.5
pydantic_core~=2.27.2
pydantic-settings~=2.7.1
typing_extensions~=4.12.2
starlette~=0.41.3
gunicorn

# Database Libraries
psycopg2-binary
sqlmodel

# SSE
sse-starlette~=2.2.1

# Testing Libraries
pytest~=8.3.4
pytest-cov~=6.0.0
pytest-dotenv~=0.5.2

# Network Libraries
aiohttp~=3.11.11
httpx~=0.28.1
redis~=5.2.1

# environment variables
python-dotenv~=1.0.1

# JWT
pyjwt~=2.10.1